Hands and feet rotations are kind of working
"Kind of". Turns out that when you rotate "Left leg"
on the Z axis, you'll be able to rotate the whole leg,
meaning that the knee, the ankle and the foot follows
the rotation.
On MMD, 左足only rotates the Thigh, and will stop at
the knee.

Some MMD modelers actually add special bones to handle
that, named 左足D. Moving the "D" version actually behave
just like on Unity.

This is why I'm using デグー for the current tests, and
will try back with アライさんmodel after, since the later
do not have 'D' bones.

Now you can move the whole leg without 'D' bones.
You just have to move the IK goals.
Since I'm converting Mecanim animations, this might be
done in the future.

But for the moment, I'm still trying to see if I can force
the conversion using FK...

Meaning that the conversion now moves the legs on some
models. However, the hips are still not moving correctly.

Signed-off-by: Voyage <voyage@miouyouyou.fr>