Almost there !
So :

* First, すべての親should actually be named 全ての親 .

That allowed me to test associating the Unity "Hips" to
全ての親.

And it worked nicely, though, I got told that this is
extremely weird and that I should associate the "Hips"
to "Center".

And then I was like "... Yeah, but if you rotate 'Hips'
in Unity, you flip the character. If you flip 'Center'
in MMD, you don't...", to which I got replied to disable
the IK.

And so

* Second, disabling the IK in a frame allow for the lower
half bones to act like simple bones, following the way
Unity move bones.

This works actually pretty fine !

When doing this, "足" move like Unity "Leg", and
"センター" move like "Hips".
So, now, the motion seems to follow Unity more closely.

I was wondering if disabling an IK was some weird VMD
functionnality unavailable to MMD users, but turns out
that the red panel on MMD actually allows you to do that.
It's just that I could not read the bones names, due to
Shift JIS striking again !

So, now, the movement around the waist still feel
extremely weird, but we're getting there, and I can now
move characters that don't have 'D' bones, meaning that
I can animate アライさん again なのだ

Signed-off-by: Voyage <voyage@miouyouyou.fr>